Create README.md
SSN - The Social Security Project

The purpose of this project is to build an application that issues and manages social security numbers to the people of Ethiopia. It is a springboot Restful webservice project written mainly in java 8. The application will have a UI built in React and an oracle database at the back, both of which are not part of this specific project for now. The project issues social security numbers, update and maintain them, produces statistical services to qualified institutions under the law, put on hold or remove them as needed. A government institution mandated by law, will make use of this api.